Tilføjet forside, forberedt til at implementere registrerings funktion.
