+++ app.js
@@ -12,6 +12,7 @@ var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
 var hesteRouter = require('./routes/heste');
 var profileRouter = require('./routes/profile');
+var loginRouter = require ('./routes/login');
 
 
 var app = express();
@@ -54,6 +55,7 @@ app.use('/', indexRouter);
 app.use('/users', usersRouter);
 app.use('/heste', hesteRouter);
 app.use('/profile', profileRouter);
+app.use('/login', loginRouter);
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
   next(createError(404));
