+++ public/stylesheets/style.css
@@ -7,13 +7,6 @@ a {
   color: #00B7FF;
 }
 
-#hestCSS {
-  border:solid;
-  background-color:pink;
-
-}
-
-
 html,
 body {
   height: 100%;
@@ -28,6 +21,8 @@ body {
   padding-bottom: 40px;
   background-color: #f5f5f5;
 }
+
+
 
 .form-signin {
   width: 100%;
@@ -57,4 +52,18 @@ body {
   margin-bottom: 10px;
   border-top-left-radius: 0;
   border-top-right-radius: 0;
+}
+
+
+.startBtn {
+  background-color: #4CAF50; /* Green */
+  border: none;
+  color: white;
+  padding: 15px 32px;
+  text-align: center;
+  text-decoration: none;
+  justify-content: center;
+  font-size: 16px;
+  margin:10px;
+  width:100%;
 }
