+++ views/index.hbs
@@ -1,15 +1,4 @@
-<!-- Page Content -->
-
-<div id="text-center">
-    <form action='/' method='post' class="form-signin">
-        <img class="mb-4" src="../../assets/brand/bootstrap-solid.svg" alt="" width="72" height="72">
-        <h1 class="h3 mb-3 font-weight-normal">Venligst login</h1>
-        <label for="inputEmail" class="sr-only">Email addresse</label>
-        <input type="text" id="inputEmail" name="logemail" class="form-control" placeholder="Email addresse" required="">
-        <label for="inputPassword" class="sr-only">Password</label>
-        <input type="password" name="logpassword" id="inputPassword" class="form-control" placeholder="Kodeord"
-            required="">
-        <input class="btn btn-lg btn-primary btn-block" type="submit" value="LOGIN NU"></input>
-        <p class="mt-5 mb-3 text-muted">&copy; 2017-2018</p>
-    </form>
+<div class="vertBtn">
+  <a href='/login'><button type="button" class="startBtn">Login</button></a>
+  <button type="button" class="startBtn">Registrer</button>
 </div>
